+++ CHANGELOG.md
@@ -72,6 +72,8 @@ without arguments starts a full interactive Python console.
   candidates, Builders+ will use list, local search and only global search if no match found.
 - Make `cmd.at_post_cmd()` always run after `cmd.func()`, even when the latter uses delays 
   with yield.
+- Add new `return_iterators` kwarg to `search_prototypes` function in order to prepare for 
+  more paginated handling of prototype returns.
 
 
 ## Evennia 0.9 (2018-2019)